requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
et-xmlfile==1.1.0
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.0
openpyxl==3.1.2
pandas==2.0.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
